+++ packages/import-bundle/demo/source-maps/start.js
@@ -1,0 +1,5 @@
+/* eslint-disable */
+import './debug.js';
+console.log('start (before breakpoint)');
+debugger;
+console.log('start (after breakpoint)');